+++ resources/views/layouts/admin/app.blade.php
@@ -40,6 +40,7 @@
     @include('layouts.warnung')
 
     <x-befehlspalette />
+    <x-keyboard />
 </body>
 
 </html>